requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.5.1
pytest==7.4.3
requests-mock==1.11.0
python-dotenv==1.0.0
flake8==6.1.0
pytest-cov